test: move constant to utils, fix types, fix tests
